requirements.txt
click==6.7
coreapi==2.1.1
Django==1.10.5
django-braces==1.10.0
django-cors-headers==2.0.0
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
httpie==0.9.9
itypes==1.1.0
oauthlib==1.1.2
Pygments==2.1.3
pyinotify==0.9.6
PyJWT==1.4.2
pytz==2016.10
requests==2.12.4
six==1.10.0
uritemplate==3.0.0
